move unit test fixtures file to folder
